rename NTS.SignalR.Server to NTS.Relay
also move it out of subdirectory as it is planned to be the only project of the Relay

https://github.com/Not-Endurance/not-timing-system/issues/336
