+++ src/Relay/NTS.Relay.csproj
@@ -14,9 +14,9 @@
   </ItemGroup>
 
   <ItemGroup>
-      <ProjectReference Include="..\..\NTS.Application\NTS.Application.csproj" />
-      <ProjectReference Include="..\..\NTS.Storage\NTS.Storage.csproj" />
-	  <ProjectReference Include="..\..\NTS.ACL\NTS.ACL.csproj" />
+	  <ProjectReference Include="..\NTS.ACL\NTS.ACL.csproj" />
+	  <ProjectReference Include="..\NTS.Storage\NTS.Storage.csproj" />
+	  <ProjectReference Include="..\NTS.Application\NTS.Application.csproj" />
   </ItemGroup>
 
 </Project>
