+++ src/Relay/Program.cs
@@ -4,9 +4,13 @@ using Not.Logging.Builder;
 using Not.ProcessUtility;
 using Not.Startup;
 using NTS.Application;
-using NTS.SignalR.Server;
-using NTS.SignalR.Server.ACL;
-using NTS.SignalR.Server.RPC;
+using NTS.Relay;
+using NTS.Relay.ACL;
+using NTS.Relay.RPC;
+using Not.ProcessUtility;
+using Not.Logging.Builder;
+using Not.Filesystem;
+using Not.Contexts;
 
 var builder = WebApplication.CreateBuilder(args);
