+++ src/Relay/Storage/ParticipationReadonlyRepository.cs
@@ -3,7 +3,7 @@ using Not.Storage.Stores;
 using NTS.Domain.Core.Aggregates;
 using NTS.Storage.Core;
 
-namespace NTS.SignalR.Server.Storage;
+namespace NTS.Relay.Storage;
 
 public class ParticipationReadonlyRepository : ReadonlySetRepository<Participation, CoreState>
 {
