+++ src/not-timing-system.sln
@@ -59,7 +59,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NTS.Judge.Tests", "..\tests
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Relay", "Relay", "{30423F4B-161F-4F62-8BA1-05C98139909B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NTS.SignalR.Server", "Relay\NTS.SignalR.Server\NTS.SignalR.Server.csproj", "{4B823C39-6D0F-4923-B65E-CEF777B6E4EF}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NTS.Relay", "Relay\NTS.Relay.csproj", "{50032FAE-AC22-4C5A-BA5C-6B2271215853}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -140,10 +140,10 @@ Global
 		{A638DD26-3D7A-48CC-9D30-9185DD05A1C6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A638DD26-3D7A-48CC-9D30-9185DD05A1C6}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A638DD26-3D7A-48CC-9D30-9185DD05A1C6}.Release|Any CPU.Build.0 = Release|Any CPU
-		{4B823C39-6D0F-4923-B65E-CEF777B6E4EF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4B823C39-6D0F-4923-B65E-CEF777B6E4EF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4B823C39-6D0F-4923-B65E-CEF777B6E4EF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4B823C39-6D0F-4923-B65E-CEF777B6E4EF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{50032FAE-AC22-4C5A-BA5C-6B2271215853}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{50032FAE-AC22-4C5A-BA5C-6B2271215853}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{50032FAE-AC22-4C5A-BA5C-6B2271215853}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{50032FAE-AC22-4C5A-BA5C-6B2271215853}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -160,7 +160,7 @@ Global
 		{A6196586-C84D-4CAE-957B-FB86B0063037} = {06233005-CC56-4846-8760-F7AFF5F72F4C}
 		{22608148-7347-4C99-9914-9A55EA74F6EB} = {06233005-CC56-4846-8760-F7AFF5F72F4C}
 		{A638DD26-3D7A-48CC-9D30-9185DD05A1C6} = {C950F8E2-9577-4169-9879-D9442F4FB631}
-		{4B823C39-6D0F-4923-B65E-CEF777B6E4EF} = {30423F4B-161F-4F62-8BA1-05C98139909B}
+		{50032FAE-AC22-4C5A-BA5C-6B2271215853} = {30423F4B-161F-4F62-8BA1-05C98139909B}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {3CB609D9-5D54-4C11-A371-DAAC8B74E430}
